chore: move the UrlsConfig to the configuration package
